refactor: 일부 함수 및 파일위치 재조정
